backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
rq==1.15.1
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0
anthropic==0.39.0
openai==1.3.7
google-generativeai==0.3.1
docker==6.1.3
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6
PyGithub==2.1.1
GitPython==3.1.40
flask
markdown
